Updated character class!
Now uses EnumMaps and should be A LOT more intuitive to use. Currently
stores Attributes (race, class, ect) and stats (dex, experience, stre,
ect) so it doesn't have everything yet, but I am going to try to add on
and feel free to add your own ways of sotring the character Data in it.
If need be we can just work with these for now since they are smaller
chunks.